+++ WTP05_YMZ/src/com/ymz/review/controller/ReviewController.java
@@ -16,7 +16,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.ymz.common.validator.ReviewValidator;
 import com.ymz.member.vo.Member;
+import com.ymz.qna.vo.QNA;
 import com.ymz.review.service.ReviewService;
 import com.ymz.review.vo.Review;
 import com.ymz.reviewreply.service.ReviewReplyService;
@@ -35,7 +37,10 @@ public class ReviewController {
 	//리뷰 등록
 	@RequestMapping(value="login/write.do", method=RequestMethod.POST)
 	public String registerReview(@ModelAttribute Review review, Errors errors, HttpSession session) throws Exception{
+		new ReviewValidator().validate(review, errors);
+		// 제목과 내용 미입력시 등록 실패
 		if(errors.hasErrors()){
+			System.out.println("모두 입력하지 않아 등록 실패");
 			return "review/review_write_form.tiles";
 		}
 		Member member = (Member)session.getAttribute("login_info");
@@ -83,24 +88,44 @@ public class ReviewController {
 	}
 	
 	//리뷰 수정(로그인시 가능)
-	@RequestMapping(value="login/modifyReview.do")
+	@RequestMapping(value="login/modifyReview.do", method=RequestMethod.POST)
 	public String modifyReview(@ModelAttribute Review review, Errors errors, HttpSession session){
-		Member member = (Member)session.getAttribute("login_info");
-		String userid = member.getId();
-		review.setMemberId(userid);
-		service.modifyReview(review); // 리뷰 수정
-		return "redirect:/review/reviewList.do";
+		new ReviewValidator().validate(review, errors);
+		// 제목과 내용 미입력시 수정 실패
+		if(errors.hasErrors()){
+			System.out.println("모두 입력하지 않아 수정 실패");
+			return "review/review_modify_form.tiles";
+		}
+		
+		// 작성자만 글을 수정할 수 있다.
+		Review r = service.getReviewByNo(review.getReviewNo());			 // 글번호로 글내용을 가져온다.
+		Member member = (Member) session.getAttribute("login_info");    // 로그인한 회원의 정보를 가져온다.
+		if(r.getMemberId().equals(member.getId())){
+			// 수정
+			String userid = member.getId();
+			review.setMemberId(userid);
+			service.modifyReview(review); // 리뷰 수정
+			return "redirect:/review/reviewList.do";
+		}
+		
+		return "view/loginInfoCheck.tiles";	// 수정 실패 페이지
 	}
 	
 	
 	//리뷰 삭제(로그인시 가능)
 	@RequestMapping("login/removeReview.do")
-	public String removeReview(@ModelAttribute Review review, Errors errors, HttpSession session){
-		Member member = (Member)session.getAttribute("login_info");
-		String userid = member.getId();
-		review.setMemberId(userid);
-		service.removeReview(review);
-		return "redirect:/review/reviewList.do";
+	public String removeReview(@ModelAttribute Review review, HttpSession session){
+		// 작성자와 관리자만 삭제할 수 있다.
+		Review r = service.getReviewByNo(review.getReviewNo());			 // 글번호로 글내용을 가져온다.
+		Member member = (Member) session.getAttribute("login_info");    // 로그인한 회원의 정보를 가져온다.
+		if(r.getMemberId().equals(member.getId()) || member.getGrade().equals("master")){
+			// 삭제
+			String userid = member.getId();
+			review.setMemberId(userid);
+			service.removeReview(review);
+			return "redirect:/review/reviewList.do";
+		}
+		return "view/loginInfoCheck.tiles";	// 삭제 실패 페이지
 	}
 	
 	// 리뷰 추천(로그인시 가능)
